added assertions for number of arguments

+++ trashbox.sh
@@ -3,7 +3,12 @@
 export TRASHBOX="${XDG_DATA_HOME:-$HOME}/.trashbox"
 
 # moves 1 or more files or directories to the trashbox
-function trashbox-put() { 
+function trashbox-put() {
+  if [ "$#" -eq 0 ] ;then
+    echo "Usage: trashbox-put <path ...>" 1>&2
+    return 1
+  fi
+
   time="$(date +%s.%N | cut -b 1-13)"
 
   mkdir -p "$TRASHBOX/$time"
@@ -12,12 +17,22 @@ function trashbox-put() {
 
 # restores a file or directory from the trashbox
 function trashbox-restore() {
+  if [ "$#" -ne 2 ] ;then
+    echo "Usage: trashbox-restore <source relative to the trashbox> <destination>" 1>&2
+    return 1
+  fi
+
   mv "$TRASHBOX/$1" "$2" &&
   (rmdir "$TRASHBOX/$(echo "$1" | cut -d / -f 1)" 2> /dev/null || true)
 }
 
 # prompts to permanently remove everything from the trashbox
 function trashbox-empty() {
+  if [ "$#" -ne 0 ] ;then
+    echo "Usage: trashbox-empty" 1>&2
+    return 1
+  fi
+
   printf "Empty trashbox? (y/N) "
 
   read answer
@@ -29,6 +44,10 @@ function trashbox-empty() {
 
 # prompts to permanently remove some files or directories from the trashbox
 function trashbox-remove() {
+  if [ "$#" -eq 0 ] ;then
+    echo "Usage: trashbox-remove <path relative to the trashbox>" 1>&2
+    return 1
+  fi
   printf "Delete the given files? (y/N) "
 
   read answer
@@ -43,6 +62,11 @@ function trashbox-remove() {
 
 # lists all files and directories in the trashbox
 function trashbox-list() {
+  if [ "$#" -ne 0 ] ;then
+    echo "Usage: trashbox-list" 1>&2
+    return 1
+  fi
+
   mkdir -p "$TRASHBOX"
   if find "$TRASHBOX" -mindepth 1 -maxdepth 1 | read; then
     find "$TRASHBOX"/*/* 2> /dev/null | cut -b $(echo "$TRASHBOX " | wc -c)-
